requirements.txt
fastapi==0.115.12
langchain_community==0.3.24
langchain_core==0.3.59
langchain_google_genai==2.1.4
langchain-groq==0.3.2
langdetect==1.0.9
langgraph==0.4.3
passlib==1.7.4
pydantic==2.11.4
PyJWT==2.9.0
python-dotenv==1.1.0
redis==5.3.0
slowapi==0.1.9
spacy==3.8.5
sqladmin==0.20.1
SQLAlchemy==2.0.24
starlette==0.46.2
textblob==0.19.0
trustcall==0.0.39
typing_extensions==4.13.2
uvicorn==0.34.2
psycopg2-binary==2.9.9
email-validator==2.1.0
wikipedia==1.4.0
pyspellchecker==0.8.2
langgraph-checkpoint-redis==0.0.5
langchain-experimental==0.3.4
langgraph-supervisor==0.0.23
langgraph-swarm==0.0.11
python-decouple==3.8
langchain-tavily==0.2.12